Refactor: Move Executor to CategoryRepository
